requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
langchain-google-genai
langchain-groq
tavily-python
plotly
google-generativeai
openai
